backend/requirements.txt
Django==4.2.10
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
celery==5.3.6
redis==5.0.1
django-storages==1.14.2
boto3==1.34.0
Pillow==10.2.0
python-decouple==3.8
psycopg2-binary==2.9.9
daphne==4.0.0
flower==2.0.1
drf-spectacular==0.27.2
